Stop clipping the Area UI
The clip rect was a very, very rough estimate that wasn't doing
anything. Because the child of an Area is usually something with a large
shadow, we had to conservatively expand the clip rect to the hardcoded
window_shadow.extrusion value, even if we're not rendering a window,
because we have no idea what will be put inside the Area.

I also changed max_rect to no longer be expanded by 32 units. I have no
idea where this number came from, and removing it could be a very bad
idea, but it seems like the clipping code has been touched a bunch of
times since that 32-unit expansion was introduced, and it smells
vestigial to me.