Stop clipping Resize
The immediate child of a Resize element may be a Frame, which wants to
expand the clip rect by its stroke radius. To let it do that, stop
clipping the Resize element entirely, because we have no idea when the
Frame will appear or if it will even be added.

An alternative solution to this conundrum would be to always draw a
Frame's outline strictly inside the space it is given, expanding
inwards. This would have more knock-on effects, but would make the box
model much sounder since a rectangle would just be a rectangle, without
the current exception of strokes being able to render outside their
bounds.